Revert "Sync Indigo devel with Hydro devel"
