+++ rtt_dynamic_reconfigure/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_dynamic_reconfigure</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>The rtt_dynamic_reconfigure package</description>
   <maintainer email="orocos-dev@lists.mech.kuleuven.be">Orocos Developers</maintainer>
