+++ rtt_ros/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_ros</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>
     This package provides an RTT plugin to add a ROS node to the RTT process,
     as well as several wrapper scripts to enable roslaunching of orocos
