+++ rtt_ros_integration/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_ros_integration</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>This stack contains all software necessary to build systems using both Orocos and ROS infrastructures</description>
   <maintainer email="orocos-dev@lists.mech.kuleuven.be">Orocos Developers</maintainer>
