+++ rtt_ros_msgs/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_ros_msgs</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>This package provides .msg and .srv files for use with the rtt_ros_integration packages.</description>
   <maintainer email="orocos-dev@lists.mech.kuleuven.be">Orocos Developers</maintainer>
