+++ rtt_rosclock/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_rosclock</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>This package provides an RTT plugin to access different time measurements on a realtime host.</description>
   <maintainer email="orocos-dev@lists.mech.kuleuven.be">Orocos Developers</maintainer>
