+++ rtt_roscomm/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_roscomm</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>
     rtt_roscomm provides the necessary template files and cmake macros for
     automatic rtt typekit and transport generation from ROS msg files
