+++ rtt_rosdeployment/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_rosdeployment</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>rtt_rosdeployment provides an RTT plugin to control an ocl deployment component over ROS service calls.</description>
   <maintainer email="orocos-dev@lists.mech.kuleuven.be">Orocos Developers</maintainer>
