+++ rtt_rospack/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_rospack</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>rtt_rospack provides an RTT plugin to use rospack to find packages in your ROS_PACKAGE_PATH</description>
   <maintainer email="orocos-dev@lists.mech.kuleuven.be">Orocos Developers</maintainer>
