+++ rtt_rosparam/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_rosparam</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>
     This package provides an RTT service and service-requester for associating RTT component properties with ROS parameters
   </description>
