+++ rtt_tf/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_tf</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>This package contains the components of the rtt_tf package</description>
   <maintainer email="orocos-dev@lists.mech.kuleuven.be">Orocos Developers</maintainer>
