+++ tests/rtt_dynamic_reconfigure_tests/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <package>
   <name>rtt_dynamic_reconfigure_tests</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>The rtt_dynamic_reconfigure_tests package</description>
   <maintainer email="orocos-dev@lists.mech.kuleuven.be">Orocos Developers</maintainer>
 
@@ -13,17 +13,14 @@
 
   <build_depend>rtt_ros</build_depend>
   <build_depend>rtt_dynamic_reconfigure</build_depend>
-  <build_depend>rtt_geometry_msgs</build_depend>
 
   <run_depend>rtt_ros</run_depend>
   <run_depend>rtt_dynamic_reconfigure</run_depend>
-  <run_depend>rtt_geometry_msgs</run_depend>
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
     <rtt_ros>
       <plugin_depend>rtt_dynamic_reconfigure</plugin_depend>
-      <plugin_depend>rtt_geometry_msgs</plugin_depend>
     </rtt_ros>
   </export>
