--- tests/rtt_dynamic_reconfigure_tests/test/test_component.cpp
@@ -1,11 +1,0 @@
-/*********************************************************************
- *
- *  Copyright (c) 2015, Intermodalics BVBA
- *  All rights reserved.
- *
- *********************************************************************/
-
-#include <rtt/Component.hpp>
-#include "test_component.hpp"
-
-ORO_CREATE_COMPONENT(DynamicReconfigureTestComponent)
