+++ typekits/rtt_kdl_conversions/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rtt_kdl_conversions</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>
     This package contains the components of the kdl_conversions package
   </description>
