+++ typekits/rtt_rosgraph_msgs/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rtt_rosgraph_msgs</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>
 
     Provides an rtt typekit for ROS rosgraph_msgs messages.
