+++ typekits/rtt_trajectory_msgs/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rtt_trajectory_msgs</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>
 
     Provides an rtt typekit for ROS trajectory_msgs messages.
