+++ typekits/rtt_visualization_msgs/package.xml
@@ -1,6 +1,6 @@
 <package>
   <name>rtt_visualization_msgs</name>
-  <version>2.7.0</version>
+  <version>2.8.1</version>
   <description>
 
     Provides an rtt typekit for ROS visualization_msgs messages.
